adds pagination component but no styles
