+++ src/App.js
@@ -1,6 +1,10 @@
+import React from "react";
+
 import "./App.css";
 import { Button } from "./stories/Button/index";
 import { HtmlTable } from "./stories/HtmlTable";
+import { Pagination } from "./stories/Pagination";
+import { default as data } from "./stories/Pagination/MOCK_DATA.json";
 
 import {
   functionalColumns,
@@ -8,6 +12,25 @@ import {
 } from "./stories/HtmlTable/HtmlTable.stories";
 
 function App() {
+  const [itemsList] = React.useState(data);
+
+  const [currentItems, setCurrentItems] = React.useState([]);
+
+  function onPageChanged(paginationData) {
+    const { currentPage, pageLimit } = paginationData;
+    const offset = (currentPage - 1) * pageLimit;
+    const currentItems = itemsList.slice(offset, offset + pageLimit);
+    setCurrentItems(currentItems);
+    // axios
+    //   .get(`/api/items?page=${currentPage}&limit=${pageLimit}`)
+    //   .then((response) => {
+    //     const currentItems = response.data;
+    //     setCurrentPage(currentPage);
+    //     setCurrentItems(currentItems);
+    //     setTotalPages(totalPages);
+    //   });
+  }
+
   return (
     <div className="wrapper">
       <div className="content">
@@ -22,6 +45,21 @@ function App() {
             stripped={true}
             headerTextAlign="center"
             stickyHeader
+          />
+        </div>
+        <div className="section">
+          <ul style={{ display: "flex", flexDirection: "column" }}>
+            {currentItems?.map((item, i) => (
+              <li key={item.name}>{item.name}</li>
+            ))}
+          </ul>
+        </div>
+        <div className="section">
+          <Pagination
+            totalItems={itemsList.length}
+            pageLimit={10}
+            pageNeighbours={1}
+            onPageChanged={onPageChanged}
           />
         </div>
       </div>
